Use shorthand notation for initial assignment to $PACKER
Copy of the github discussion here:

== rasa

This could be simplfied to:

PACKER ?= packer

== Simon Kohlmeyer
My make-foo is pretty basic, so I went to look that up[1]

It seems that PACKER ?= packer is equivalent to PACKER = packer if PACKER is not
set, which is semantically different from PACKER := packer. The latter is
a simply expanded variable, which means that if $packer is defined, it will
still evaluate to "packer", while the shorthand notation will evaluate to the
value of $packer.
I left this here to document my findings, but as I don't
think it makes any difference here and imo the shorthand notation looks nicer,
I'll change it.

[1] https://www.gnu.org/software/make/manual/make.html#Setting